+++ src/libtrx/game/camera/photo_mode.c
@@ -164,8 +164,29 @@ static void M_ClampCameraPos(void)
     g_Camera.target.z += (g_Camera.pos.z - prev_cam_pos.z);
 }
 
+static bool M_CameraOutsideRoom(const XYZ_32 pos, const int16_t room_num)
+{
+    const BOUNDS_32 bounds = Room_GetRoomBounds(room_num);
+    return (
+        pos.x < bounds.min.x + WALL_L || pos.y < bounds.min.y
+        || pos.z < bounds.min.z + WALL_L || pos.x > bounds.max.x - WALL_L
+        || pos.y > bounds.max.y || pos.z > bounds.max.z - WALL_L);
+}
+
 static void M_UpdateCameraRooms(void)
 {
+    Room_GetSector(
+        g_Camera.pos.x, g_Camera.pos.y, g_Camera.pos.z, &g_Camera.pos.room_num);
+    Room_GetSector(
+        g_Camera.target.x, g_Camera.target.y, g_Camera.target.z,
+        &g_Camera.target.room_num);
+
+    if (!M_CameraOutsideRoom(g_Camera.pos.pos, g_Camera.pos.room_num)
+        && !M_CameraOutsideRoom(
+            g_Camera.target.pos, g_Camera.target.room_num)) {
+        return;
+    }
+
     const int16_t pos_room_num =
         Room_GetIndexFromPos(g_Camera.pos.x, g_Camera.pos.y, g_Camera.pos.z);
     const int16_t tar_room_num = Room_GetIndexFromPos(