+++ docs/tr2/CHANGELOG.md
@@ -63,6 +63,7 @@
 - fixed caustics stopping after spending roughly 12 minutes in a level (#4109, regression from 1.4)
 - fixed Lara being able to push blocks through toggle opacity 1 portals (#4129, regression from 1.5)
 - fixed pistols disappearing from Lara's holsters in the cutscene following The Great Wall (#4145, regression from 0.9)
+- fixed photo mode camera clipping through overlapping rooms
 
 ## [1.5.1](https://github.com/LostArtefacts/TRX/compare/tr2-1.5...tr2-1.5.1) - 2025-10-10
 - changed the examine dialog to be usable with non-puzzle items (#4009)